Move _CRT_SECURE_NO_DEPRECATE form zlib to top level, and use the newer form.
This is generally preferable when compiling third party libs, so it
doesn't bite later. I've also converted it from the older NO_DEPRECATE
to the newer NO_WARNINGS form.

For those not in the know, it basically tells MSVC to STFU about telling
you to use Microsoft's whatever_s() functions in place of the standard
whatever() functions.