Adapt sqlite3 module to MSVC; change build options
Added is the .def file from the corrisponding
sqlite-dll-win32-x86-%VERSION%.zip file, that is 'supposed' to match what is in
the sqlite-amalgamation-%VERSION%.zip file, we are using for source input.

The column metadata and rtree parts of the API are usually disabled. But since
they are apart of the .def file for MSVC, may as well enable them all the
way...